% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.penyw.R
\name{ar.penyw}
\alias{ar.penyw}
\title{ar.penyw}
\usage{
ar.penyw(
  x,
  aic = TRUE,
  order.max = NULL,
  lh = NULL,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  ...
)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{aic}{logical, If \code{TRUE} then the Akaike Information Criterion is used to choose the order of the autoregressive model. If \code{FALSE}, the model of order \code{order.max} is fitted.}

\item{order.max}{maximum order (or order) of model to fit.  Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)}
where \eqn{N} is the number of non-missing observations except for \code{method="mle"} where it is the minimum of this quantity and 12.  \code{nser} is the number of series.}

\item{na.action}{function to be called to handle missing values. Currently, via \code{na.action=na.pass}, only (penalized) Yule-Walker methods can handle missing values which must be consistent within a time point; either all variables must be missing or none.}

\item{demean}{logical. Should a mean be estimated and subtracted before correlations are calculated?}

\item{series}{names for the series. Defaults to \code{deparse1(substitute(x))}.}

\item{...}{additional arguments for \code{ar} function.}
}
\value{
An object of penalized ar model fit with the following elements:
\describe{
\item{\code{order}}{The order of the fitted model. This is chosen by minimizing the AIC if 'AIC=TRUE', otherwise it is 'order.max'.}
\item{\code{penar}}{Estimated penalized autorregression coefficients for the fitted model.}
\item{\code{aic}}{The vector of AIC values.}
\item{\code{n.used}}{The number of observations in the time series.}
\item{\code{order.max}}{The value of the 'order.max' argument.}
\item{\code{call}}{The matched call.}
}
}
\description{
Fit an autoregressive time series model to the data using penalized correlation estimation, by default selecting the order by AIC.
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Example for penalized ar model fit
ar.penyw(data)
}
}
\references{
Gallagher, C., Killick, R., Tan, X. (2024+) Penalized M-estimation 
for Autocorrelation. \emph{Submitted.}
}
\keyword{internal}
