% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{ar}
\alias{ar}
\title{ar}
\usage{
ar(
  x,
  aic = TRUE,
  order.max = NULL,
  method = c("penyw", "yule-walker", "burg", "ols", "mle", "yw"),
  na.action = na.fail,
  series = deparse1(substitute(x)),
  ...
)
}
\arguments{
\item{x}{a univariate numeric time series.}

\item{aic}{'logical', If 'TRUE' then the Akaike Information Criterion is used to choose the order of the autoregressive model. If FALSE, the model of order order.max is fitted.}

\item{order.max}{maximum lag at which to calculate the penalized/sample PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{method}{character string specifying the method to fit the regular autoregressive time series model. Default to "yule-walker".}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{series}{names for the series. Defaults to \code{deparse1(substitute(x))}.}

\item{...}{additional arguments for \code{ar} function.}
}
\value{
An object of penalized ar model fit with the following elements:
\describe{
\item{\code{order}}{The order of the fitted model. This is chosen by minimizing the AIC if 'AIC=TRUE', otherwise it is 'order.max'.}
\item{\code{ar}}{Estimated penalized autorregression coefficients for the fitted model.}
\item{\code{aic}}{The vector of AIC values.}
\item{\code{n.used}}{The number of observations in the time series.}
\item{\code{order.max}}{The value of the 'order.max' argument.}
\item{\code{call}}{The matched call.}
}
}
\description{
Fit an autoregressive time series model to the data using penalized correlation estimation, by default selecting the order by AIC.
series = deparse1(substitute(x)),
}
\examples{
\dontrun{
set.seed(1234)
data <- arima.sim(n=100, model=list(ar=0.5))

ar(data) # penalized ar model fit
ar(data,method="yw") # default stats::ar() estimate

}
}
