% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.acf.R
\name{auto.acf}
\alias{auto.acf}
\title{auto.acf}
\usage{
auto.acf(
  x,
  lag.max = NULL,
  order.max = 2,
  plot = TRUE,
  na.action = na.fail,
  demean = TRUE,
  penalized = TRUE,
  lh = NULL,
  ...
)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the penalized/sample ACF/PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{order.max}{the maximum AR order to invert, defaults to 2}

\item{plot}{'logical'. If 'TRUE' (the default) the penalized/sample ACF/PACF is plotted.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{demean}{'logical'. Should a mean be estimated during estimating.}

\item{penalized}{'logical'. If 'TRUE' (the default) the penalized ACF/PACF is computed; if 'FALSE' the sample ACF/PACF is computed.}

\item{lh}{sequence of threshold values across 1:lag.max. Could be a single value (repeated for all 1:lag.max), a single vector of length lag.max (repeated for all nser), or a lag.max x nser matrix. Default is NULL and thus data driven.}

\item{...}{additional arguments for specific methods.}
}
\value{
An object of penalized/sample ACF estimation with the following elements:
\describe{
\item{\code{acf}}{A max.lag x nseries x nseries array containing the estimated penalized acf.}
\item{\code{type}}{Character vector returning the type argument requested.}
\item{\code{n.used}}{Numeric of the number of points used for estimation after na.action has been applied.}
\item{\code{lag}}{A max.lag x nseries x nseries array containing the lags at which the acf/pacf is estimated.}
\item{\code{series}}{The name of the time series.}
\item{\code{snames}}{The series names for a multivariate time series.}
\item{\code{lh}}{The lh used in the estimation for reproducibility.}
\item{\code{penalized}}{Logical indicating if the acf/pacf returned is penalized.}
\item{\code{estimate}}{Type of estimate provided "auto" for reproducibility.}
}
}
\description{
The Penalized (Partial) Autocorrelation Function (ACF/PACF) Estimation
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Examples for penalized ACF/PACF and sample ACF/PACF
acf(data)
acf(data, penalized = FALSE)
acf(data, type ="partial")
acf(data, type ="partial", penalized = FALSE)

x1 <- arima.sim(n=100, model=list(ar=0.5))
x2 <- arima.sim(n=100, model=list(ar=0.1))
x3 <- arima.sim(n=100, model=list(ar=0.9))
x <- cbind(x1, x2, x3)

acf(x)
acf(x, penalized = FALSE)
acf(x, type ="partial")
acf(x, type ="partial", penalized = FALSE)
}
}
\references{
Gallagher, C., Killick, R., Tan, X. (2024+) Penalized M-estimation 
for Autocorrelation. \emph{Submitted.}
}
