% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected.R
\name{corrected}
\alias{corrected}
\title{corrected}
\usage{
corrected(
  x,
  lag.max = NULL,
  type = c("correlation", "covariance", "partial"),
  na.action = na.fail,
  demean = TRUE,
  lh = NULL,
  target = apply(nserIndexM, MARGIN = 1, FUN = function(i) {
     ind = (abs(tmpacf[, i,
    i]) > lh[, i])
     target = b[, i] * ind
     return(target)
 }),
  lambda = apply(nserIndexM, MARGIN = 1, FUN = function(i) {
     ind = (abs(tmpacf[, i,
    i]) > lh[, i])
     lambda = (!ind) * h * 10 * log10(sampleT) * lh[, i] * (lh[, i] -
    abs(tmpacf[, i, i]))/abs(tmpacf[, i, i])^{
         3
     } + (ind) * (abs(tmpacf[,
    i, i]) - lh[, i]) * (1 - lh[, i])/(1 - abs(tmpacf[, i, i]))^2 * 10 * log10(sampleT)
 
       return(lambda)
 }),
  ...
)
}
\arguments{
\item{x}{a univariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the penalized/sample PACF. Defaults to \eqn{10log_{10}(N)} where N is the number of non-missing observations.}

\item{type}{'character'. "correlation", "covariance" or "partial".}

\item{na.action}{function to be called to handle missing values. Default is na.fail, na.pass can be used.}

\item{demean}{'logical'. If 'TRUE' (the default), mean(x) is removed prior to estimation.}

\item{lh}{vector of length 1 (value used for all lags), or length lag.max. Default uses formula in the description.}

\item{...}{additional arguments passed to plotting.}
}
\value{
An object of type acf with the following elements:
\describe{
\item{\code{acf}}{A max.lag x nseries x nseries array containing the estimated penalized acf/pacf.}
\item{\code{type}}{Character vector returning the type argument requested.}
\item{\code{n.used}}{Numeric of the number of points used for estimation after na.action has been applied.}
\item{\code{lag}}{A max.lag x nseries x nseries array containing the lags at which the acf/pacf is estimated.}
\item{\code{series}}{The name of the time series.}
\item{\code{snames}}{The series names for a multivariate time series.}
\item{\code{penalized}}{Logical indicating if the acf/pacf returned is penalized.}
}
}
\description{
Calculated the corrected (partial) autocorrelation function
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Example for penalized acf estimation
corrected(data)
}
}
\references{
Gallagher, C., Killick, R., Tan, X. (2024+) Penalized M-estimation 
for Autocorrelation. \emph{Submitted.}
}
\keyword{internal}
