% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLpencoef.R
\name{DLpencoef}
\alias{DLpencoef}
\title{DLpencoef}
\usage{
DLpencoef(x, lag.max = NULL, ...)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the coefficients. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{...}{additional arguments for penalized PACF estimation.}
}
\value{
A coefficients vector to estimate a penalized autocorrelation function which is positive definite.
}
\description{
Compute coefficients via the Durbin Levinson algorithm using the penalized partial autocorrelation function (PACF) estimation to obtain an penalized autocorrelation function (ACF) estimation which is positive definite.
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

DLpencoef(data)
DLpencoef(data, lag.max=10)
}
}
