% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getaic.R
\name{getaic}
\alias{getaic}
\title{getaic}
\usage{
getaic(x, lag.max = NULL, na.action = na.fail, ...)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the penalized PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{...}{additional arguments for penalized PACF estimation.}
}
\value{
A vector of AIC calaulated using penalized partial autocorrelation function estimation.
}
\description{
Calculate the value of the Akaike Information Criterion (AIC) based on penalized PACF, which is used to choose the order of the autoregressive model.
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

getaic(data)
}
}
