% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNDpenacf.R
\name{NNDpenacf}
\alias{NNDpenacf}
\title{NNDpenacf}
\usage{
NNDpenacf(x, penacf, lag.max = NULL, ...)
}
\arguments{
\item{x}{a univariate numeric time series or a numeric vector.}

\item{penacf}{vector of penalized ACF with posdef = "orginal".}

\item{lag.max}{maximum lag at which to calculate the penalized ACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{...}{additional arguments for penalized ACF/PACF estimation.}
}
\value{
An object of penalized positive definite ACF estimation, which is a list with the following elements:
\describe{
\item{\code{rhotilde.NND}}{A vector containing the estimated penalized ACF which is Nonnegative definite.}
\item{\code{lag.max}}{The value of the 'lag.max' argument.}
}
}
\description{
Calculte Nonnegative definifite (NND) penalized estimator of autocorrelation function (ACF).
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

penacf1 <- penacf(data, posdef="origin")
NNDpenacf(data, penacf1)
}
}
