% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penacf.R
\name{penacf}
\alias{penacf}
\title{penacf}
\usage{
penacf(
  x,
  lag.max = NULL,
  C = NULL,
  type = c("correlation", "partial"),
  penalized = TRUE,
  posdef = c("origin", "DL", "NND"),
  print = TRUE,
  plot = TRUE,
  series = NULL,
  na.action = na.fail,
  demean = TRUE,
  ...
)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the penalized/sample ACF/PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{C}{a positive real number for l_h}

\item{type}{type of acf to be computed. Allow values are "correlation" (the default) or "partial".}

\item{penalized}{'logical'. If 'TRUE' (the default) the penalized ACF/PACF is computed; if 'FALSE' the sample ACF/PACF is computed.}

\item{posdef}{the positive definite (two methods) penalized ACF is computed. (default is "origin") Allow values are "origin", "DL" or "NND".}

\item{print}{'logical'. If 'TRUE' (the default) the penalized/sample ACF/PACF is printed.}

\item{plot}{'logical'. If 'TRUE' (the default) the penalized/sample ACF/PACF is plotted.}

\item{series}{names for the series. Defaults to 'deparse(substitute(x))'.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{demean}{'logical'. Should a mean be estimated during estimating.}

\item{...}{additional arguments for specific methods.}
}
\value{
An object of penalized/sample ACF/PACF estimation with the following elements:
\describe{
\item{\code{acf}}{An array containing the estimated penalized/sample ACF/PACF.}
\item{\code{lag}}{An array containing the lags at which the ACF/PACF is estimated.}
\item{\code{n.used}}{The number of observation in the time series.}
\item{\code{series}}{The name of the time series.}
\item{\code{snames}}{The series names for a multivariate time series.}
}
}
\description{
The Penalized (Partial) Autocorrelation Function (ACF/PACF) Estimation
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Examples for penalized ACF/PACF and sample ACF/PACF
penacf(data)
penacf(data, penalized = FALSE)
penacf(data, type ="partial")
penacf(data, type ="partial", penalized = FALSE)

x1 <- arima.sim(n=100, model=list(ar=0.5))
x2 <- arima.sim(n=100, model=list(ar=0.1))
x3 <- arima.sim(n=100, model=list(ar=0.9))
x <- cbind(x1, x2, x3)

penacf(x)
penacf(x, penalized = FALSE)
penacf(x, type ="partial")
penacf(x, type ="partial", penalized = FALSE)
}
}
