% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penacfauto.R
\name{penacfauto}
\alias{penacfauto}
\title{penacfauto}
\usage{
penacfauto(x, lag.max = NULL, na.action = na.fail, ...)
}
\arguments{
\item{x}{a univariate numeric time series or a numeric vector.}

\item{lag.max}{maximum lag at which to calculate the penalized ACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{...}{additional arguments for penalized ACF estimation.}
}
\value{
An object of penalized ACF estimation, which is a list with the following elements:
\describe{
\item{\code{orders}}{The vector of selected orders from \code{Penar}.}
\item{\code{penacf}}{The penalized ACF calculated by auto algorithm.}
}
}
\description{
An automatic algorithm to estimate the ACF and PACF: (1) select the order p from 'Penar'; (2) if p less than 3, the penalized ACF caluculated by inverting the penalized PACF estimator (posdef="DL"), otherwise use the original penalized ACF estimator.
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

penacfauto(data)
}
}
