% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penar.R
\name{penar}
\alias{penar}
\title{penar}
\usage{
penar(
  x,
  order.max = NULL,
  penalized = TRUE,
  AIC = TRUE,
  na.action = na.fail,
  arprint = TRUE,
  method = c("yule-walker", "burg", "ols", "mle", "yw"),
  series = deparse(substitute(x)),
  ...
)
}
\arguments{
\item{x}{a univariate numeric time series.}

\item{order.max}{maximum lag at which to calculate the penalized/sample PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{penalized}{'logical'. If 'TRUE' (the default) the penalized autoregressive model is fitted; if 'FALSE' the \code{ar} is fitted and \code{method} can be selected.}

\item{AIC}{'logical', If 'TRUE' then the Akaike Information Criterion is used to choose the order of the autoregressive model. If FALSE, the model of order order.max is fitted.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{arprint}{'logical'. If 'TRUE' (the default) the fitted penalized ar model is printed.}

\item{method}{character string specifying the method to fit the regular autoregressive time series model. Default to "yule-walker".}

\item{...}{additional arguments for \code{ar} function.}
}
\value{
An object of penalized ar model fit with the following elements:
\describe{
\item{\code{order}}{The order of the fitted model. This is chosen by minimizing the AIC if 'AIC=TRUE', otherwise it is 'order.max'.}
\item{\code{penar}}{Estimated penalized autorregression coefficients for the fitted model.}
\item{\code{aic}}{The vector of AIC values.}
\item{\code{n.used}}{The number of observations in the time series.}
\item{\code{order.max}}{The value of the 'order.max' argument.}
\item{\code{call}}{The matched call.}
}
}
\description{
Fit an autoregressive time series model to the data using penalized correlation estimation, by default selecting the order by AIC.
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Example for penalized ar model fit
penar(data)
}
}
