% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pencorrplot.R
\name{pencorrplot}
\alias{pencorrplot}
\title{pencorrplot}
\usage{
pencorrplot(
  x,
  ci = 0.95,
  ci.type = c("white", "ma"),
  type = "h",
  xlab = "Lag",
  ylab = NULL,
  ylim = NULL,
  main = NULL,
  ci.col = "blue",
  ...
)
}
\arguments{
\item{x}{an object of "acf.out" or "pacf.out".}

\item{...}{additional arguments for penalized PACF estimation.}

\item{lag.max}{maximum lag at which to calculate the coefficients. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}
}
\value{
plot of the penalized acf or pacf estimation
}
\description{
Plot method for "Penacf" and "Penpacf"
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

pencorrplot(penacf(data))

}
}
