% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pencorrprint.R
\name{pencorrprint}
\alias{pencorrprint}
\title{pencorrprint}
\usage{
pencorrprint(x, digits = 3L, ...)
}
\arguments{
\item{x}{an object of "acf.out" or "pacf.out".}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
list of results for penalized acf or pacf
}
\description{
Print of objects from "\code{penacf}" and "\code{penpacf}"
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))
pencorrprint(penacf(data))
}
}
