% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penpacf.R
\name{penpacf}
\alias{penpacf}
\title{penpacf}
\usage{
penpacf(
  x,
  lag.max = NULL,
  C = NULL,
  penalized = TRUE,
  print = TRUE,
  plot = TRUE,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  ...
)
}
\arguments{
\item{x}{a univariate or multivariate numeric time series.}

\item{lag.max}{maximum lag at which to calculate the penalized/sample PACF. Defaults to the smaller of \eqn{N-1} and \eqn{10log_{10}(N/nser)} where N is the number of non-missing observations and nser is the number of series.}

\item{C}{a positive real number for l_h}

\item{penalized}{'logical'. If 'TRUE' (the default) the penalized PACF is computed; if 'FALSE' the sample PACF is computed.}

\item{print}{'logical'. If 'TRUE' (the default) the penalized/sample PACF is printed.}

\item{plot}{'logical'. If 'TRUE' (the default) the penalized/sample PACF is plotted.}

\item{na.action}{function to be called to handle missing values. 'na.pass' can be used.}

\item{demean}{'logical'. Should a mean be estimated during estimating.}

\item{series}{names for the series. Defaults to 'deparse(substitute(x))'.}

\item{...}{additional arguments for specific methods.}
}
\value{
An object of penalized/sample PACF estimation, which is a list with the following elements:
\describe{
\item{\code{acf}}{An array containing the estimated penalized/sample PACF.}
\item{\code{lag}}{An array containing the lags at which the PACF is estimated.}
\item{\code{n.used}}{The number of observation in the time series.}
\item{\code{series}}{The name of the time series.}
\item{\code{snames}}{The series names for a multivariate time series.}
}
}
\description{
The Penalized Partial Autocorrelation Function (PACF) Estimation
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.5))

# Examples for penalized PACF and sample PACF
penpacf(data)
penpacf(data, penalized = FALSE)
}
}
