% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandtap.R
\name{bandtap}
\alias{bandtap}
\title{bandtap}
\usage{
bandtap(acf, sampleT, ...)
}
\arguments{
\item{acf}{incoming acf value}

\item{...}{additional arguments passed to plotting.}

\item{n}{incoming length of x from acf funtion}
}
\description{
Calculated the corrected (partial) autocorrelation function
}
\references{
Gallagher, C., Killick, R., Tan, X. (2024+) Penalized M-estimation 
for Autocorrelation. \emph{Submitted.}
McMurray & Politis (2010) forecast package for CI
}
